+++ helm-app-operator/pkg/helm/release/helm.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm
+package release
 
 import (
 	"bytes"
