+++ helm-app-operator/pkg/helm/release/helm_test.go
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm
+package release
